+++ Cargo.toml
@@ -6,8 +6,7 @@ members = [
     "crypto",
     "rmc",
     "mock",
-]
 
-exclude = [
-    "fuzz",
+    # fuzzing
+    "fuzz"
 ]
