+++ fuzz/Cargo.toml
@@ -1,29 +1,25 @@
 [package]
-name = "alephbft-fuzz"
-version = "0.0.1"
+name = "aleph-bft-fuzz"
+version = "0.0.0"
+edition = "2021"
 authors = ["Cardinal Cryptography"]
 publish = false
-edition = "2018"
 
 [package.metadata]
 cargo-fuzz = true
 
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
 [dependencies]
+aleph-bft = { path = "../consensus", version = "*" }
+aleph-bft-mock = { path = "../mock", version = "*" }
 async-trait = "0.1"
 codec = { package = "parity-scale-codec", version = "2", default-features = false, features = ["derive", "std"] }
 futures = "0.3"
 futures-timer = "3.0"
 log = "0.4"
+parking_lot = "0.12"
 structopt = "0.3.21"
 tokio = { version = "1.6.1", features = ["rt"] }
-parking_lot = "0.11"
 
-aleph-bft = { path = ".." }
-aleph-mock = { path = "../mock" }
 libfuzzer-sys = { version = "0.4.2", optional = true }
 afl = { version = "0.10.1", optional = true }
