+++ fuzz/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod fuzz;
+pub mod spawner;
